Move and finalize RedditMemes Connector
