+++ src/connectors/reddit/redditMemes.ts
@@ -1,11 +1,11 @@
-/* eslint-disable import/extensions, import/no-unresolved, no-continue */
-import Axios from 'axios';
+/* eslint-disable import/extensions, import/no-unresolved */
 import { LocalStorage } from 'node-localstorage';
-import Preferences from '../preferences';
+import Axios from 'axios';
+import Preferences from '../../preferences';
 
 const localStorage = new LocalStorage('./localstorage/uc_entertainment');
 
-export default class Entertainment {
+export default class RedditMemes {
   static async getMeme() {
     const postJson = await Axios.get(`https://www.reddit.com/${Preferences.get('redditMemes', 'subName')}.json?limit=50`).then((res) => {
       const posts = res.data.data.children.map((post) => post.data);
@@ -13,22 +13,16 @@ export default class Entertainment {
       do {
         const post = posts[0];
 
-        /*
-        Skip posts if they are:
-          already sent
-          sticky (usually mod posts)
-          not an image
-        */
+        // Skip posts if they are already sent, sticky (usually mod posts) or not an image
         if (post.stickied || post.id === localStorage.getItem(post.id) || post.post_hint !== 'image') {
           posts.shift();
-          continue;
+        } else {
+          // Found meme to send
+          // Remember the meme to not send it again
+          localStorage.setItem(post.id, post.id);
+
+          return post;
         }
-
-        // Found meme to send
-        // Remember the meme to not send it again
-        localStorage.setItem(post.id, post.id);
-
-        return post;
       } while (posts.length > 0);
       return undefined;
     });
