+++ src/usecases/entertainment/entertainment.test.ts
